+++ src/main/java/fr/alasdiablo/janoeo/init/AllBlocks.java
@@ -1,0 +1,10 @@
+package fr.alasdiablo.janoeo.init;
+
+public final class AllBlocks {
+    // TODO
+    public static final class Ores {
+        public static final class Nether {
+            // TODO
+        }
+    }
+}