+++ src/main/java/fr/alasdiablo/janoeo/init/AllItems.java
@@ -1,0 +1,42 @@
+package fr.alasdiablo.janoeo.init;
+
+import net.minecraft.item.Item;
+
+public final class AllItems {
+
+    public static final class Dusts {
+        public static final Item ALUMINIUM = DustsItems.ALUMINIUM_DUST;
+        public static final Item COAL = DustsItems.COAL_DUST;
+        public static final Item COPPER = DustsItems.COPPER_DUST;
+        public static final Item DIAMOND = DustsItems.DIAMOND_DUST;
+        public static final Item EMERALD = DustsItems.EMERALD_DUST;
+        public static final Item GOLD = DustsItems.GOLD_DUST;
+        public static final Item IRON = DustsItems.IRON_DUST;
+        public static final Item LAPIS = DustsItems.LAPIS_DUST;
+        public static final Item LEAD = DustsItems.LEAD_DUST;
+        public static final Item NICKEL = DustsItems.NICKEL_DUST;
+        public static final Item PLATINUM = DustsItems.PLATINUM_DUST;
+        public static final Item RUBY = DustsItems.RUBY_DUST;
+        public static final Item SAPPHIRE = DustsItems.SAPPHIRE_DUST;
+        public static final Item SILVER = DustsItems.SILVER_DUST;
+        public static final Item TIN = DustsItems.TIN_DUST;
+        public static final Item URANIUM = DustsItems.URANIUM_DUST;
+    }
+
+    public static final class Ingots {
+        public static final Item ALUMINIUM = IngotsItems.ALUMINIUM_INGOT;
+        public static final Item COPPER = IngotsItems.COPPER_INGOT;
+        public static final Item LEAD = IngotsItems.LEAD_INGOT;
+        public static final Item OSMIUM = IngotsItems.OSMIUM_INGOT;
+        public static final Item SILVER = IngotsItems.SILVER_INGOT;
+        public static final Item TIN = IngotsItems.TIN_INGOT;
+        public static final Item URANIUM = IngotsItems.URANIUM_INGOT;
+        public static final Item ZINC = IngotsItems.ZINC_INGOT;
+    }
+
+    public static final class Gems {
+        public static final Item RUBY = GemsItems.RUBY;
+        public static final Item SAPPHIRE = GemsItems.SAPPHIRE;
+        public static final Item AMETHYST = GemsItems.AMETHYST;
+    }
+}